cutscene fix/revert, gf adjust, move gc to state logic
Cutscenes worked fine, I was just stupid. I added a destroyOnEnd flag to cutscenes so they can remain and continue to update after all the events have passed. GF's hair blowing animations have been adjusted to have consistent hair timing between frames. States now garbage collect after running create() instead of relying on a transition to do it for them.

+++ source/cutscenes/ScriptedCutscene.hx
@@ -11,7 +11,16 @@ import flixel.FlxBasic;
 class ScriptedCutscene extends FlxBasic
 {
 
+	/*
+	 *	The amount of time elapsed since the cutscene started playing.
+	 */
 	public var totalTime:Float = 0;
+
+	/*
+	 *	Automatically destroys the cutscene object after all events are finished.
+	 *	Can be disabled for stuff like video events where you want to be able to keep checking to skip after the video event happened.
+	 */
+	public var destroyOnEnd:Bool = true;
 
 	public var started(get, never):Bool;
 	var __started:Bool = false;
@@ -28,7 +37,9 @@ class ScriptedCutscene extends FlxBasic
 				totalTime += elapsed;
 			}
 			else{
-				Utils.destroyWhenAvailable(this);
+				if(destroyOnEnd){
+					Utils.destroyWhenAvailable(this);
+				}
 			}
 			
 			for(event in events){
